SPT-1969: Renamed MessageProcessorLambda part 1 (new resources)
Duplicate MessageProcessorLambda to InboundTxmaQueueMessageHandler,
which now follows our preferred naming strategy. The event source
mapping will be switched to point to the new function in a future PR.
