+++ src/handlers/inbound-txma-queue-message-handler/tests/inbound-txma-queue-message-handler.test.ts
@@ -1,5 +1,5 @@
 import { SQSEvent, SQSRecord } from "aws-lambda";
-import { handler } from "../ais-message-handler";
+import { handler } from "../inbound-txma-queue-message-handler";
 import { mockClient } from "aws-sdk-client-mock";
 import { SecretsManagerClient, GetSecretValueCommand, GetSecretValueResponse } from "@aws-sdk/client-secrets-manager";
 import {
